+++ dbt/task/compile.py
@@ -3,7 +3,7 @@ from __future__ import print_function
 from dbt.logger import GLOBAL_LOGGER as logger
 from dbt.runner import RunManager
 from dbt.node_runners import CompileRunner
-from dbt.utils import NodeType
+from dbt.node_types import NodeType
 import dbt.ui.printer
